+++ src/Glob/GlobRunner.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace Glob
+namespace GlobExpressions
 {
     //class GlobRunner
     //{
@@ -11,7 +11,6 @@ namespace Glob
     //    private int _position;
     //    public bool Scan(GlobNode node, string input)
     //    {
-            
     //        if (node.Type != GlobNodeType.Tree)
     //            throw new InvalidOperationException();
 
@@ -37,7 +36,6 @@ namespace Glob
 
     //    private void ProcessRoot(GlobNode root)
     //    {
-            
     //    }
     //}
 }