+++ test/Glob.Tests/GlobTests.cs
@@ -2,7 +2,7 @@
 using System.Text;
 using Xunit;
 
-namespace Glob.Tests
+namespace GlobExpressions.Tests
 {
     public class GlobTests
     {