+++ ChangeLog.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- All types are now located in the GlobExpressions namespace instead of Glob
+
 ### Fixed
 - Fix bug where not all files are returned when using GlobExtensions